finalize favorites in component
