+++ frontend/assets/interfaces/Page.tsx
@@ -4,3 +4,8 @@ export interface HomePageProps {
   faves: number[];
   setFaves: Dispatch<SetStateAction<number[]>>;
 }
+
+export interface ExportPageProps {
+  faves: number[];
+  setFaves: Dispatch<SetStateAction<number[]>>;
+}
