+++ frontend/components/TimeFilters.tsx
@@ -1,4 +1,3 @@
-import { Dispatch, SetStateAction } from "react";
 import AutocompleteTimeComponent from "@/components/UI/AutoCompleteTime";
 import getQuarterHours from "@/helpers/getQuarterHours";
