+++ frontend/pages/export/index.tsx
@@ -1,0 +1,7 @@
+export default function ExportPage () {
+  return (
+    <>
+      Export
+    </>
+  )
+}
